Update services.md
Each of the sections has sentences that end with periods. Added a period for the sentence under "Embedding microservice" to maintain consistency with the other sections. I don't think this warrants a README.md update, but if it does then I'll be happy to add a patch change as per SemVar standards for such patches.